back/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Cors
pymysql
cryptography
bcrypt
google-generativeai==0.3.2
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.26
langchain-community>=0.0.13
langchain-core>=0.1.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
python-dotenv==1.0.0
duckduckgo-search==4.1.1
